Only consider "device" interfaces in address autodetection on Linux
- This automatically rules out bridges and other non system
  created interfaces
